State the paid-POST-route preamble once instead of four times
SonarCloud's gate stayed at exactly 3.4% duplication on new code after the last
two blocks were removed, which was the useful signal: neither was the one.
`sonarcloud.io` is blocked by this environment's egress policy, so the detector
was reproduced locally — token-based, literals normalised, 100-token minimum,
each duplicated block attributed to the lines this branch actually added. It
named the block at once: `api/story/continue.ts` lines 15-55 against both
`api/story/generate.ts` and `api/image/generate.ts`, 144 tokens, covering all 18
of this branch's added lines in that file. 18 of 518 new lines is 3.47% — the
number the gate reported.

So the duplication was the fix. Giving `/api/story/continue` the correlation id
it was missing made its preamble the fourth verbatim copy of one the other three
paid POST routes already shared: correlation id, `X-Request-ID`, CORS, method,
access control, in that order, each step depending on the last.

The gate was right, and right about the shape of the fix. That sequence is a
rule about paid routes rather than a detail of any one of them, and a rule
stated four times is a rule that can be changed in three places — which is how
`/api/story/continue` came to be missing the correlation id to begin with.
`api/_lib/http/postRoutePreamble.ts` states it once and all four routes call it;
the four handlers lose 133 lines between them and gain 28.

`/api/export/save` picks up the 405 log line the other three already wrote,
which is the same consistency this change is about.

Reproduced at 0.0% duplication on new code locally. `tests/api-access-control.test.ts`,
`tests/request-correlation-id.test.ts`, and `tests/express-api-routes.test.ts`
all still pass, along with the rest of the suite and the Angular build.
