+++ PR70_RECOVERY_CHANGELOG.md
@@ -22,7 +22,9 @@ Self-review:
 
 Review follow-up (same slice):
 
-- SonarCloud's quality gate failed at 3.4% duplication on new code against a 3% ceiling, and it was this slice's duplication rather than a standing failure — the gate passed on #242 and #243. Two blocks: `trimTrailingSeparators` was a copy of `ExportService`'s differing only in which character it trims, and is now a single `slice`, since the split-and-join leaves single separators and the cap can strand at most one; and the five-line `assert` helper each new test file re-declared now lives in `tests/assert.ts` for new tests to import. The forty-odd existing copies are left alone — rewriting every test in the repository is its own commit. That helper is the most duplicated block here, and on a small change that adds a test file those five lines are new duplicated code, enough to fail the gate on their own.
+- SonarCloud's quality gate failed at 3.4% duplication on new code against a 3% ceiling, and it was this slice's duplication rather than a standing failure — the gate passed on #242 and #243. Two blocks were removed on the first attempt, and neither was the one: `trimTrailingSeparators` was a copy of `ExportService`'s differing only in which character it trims, and is now a single `slice`, since the split-and-join leaves single separators and the cap can strand at most one; and the five-line `assert` helper each new test file re-declared now lives in `tests/assert.ts` for new tests to import. The forty-odd existing copies are left alone — rewriting every test in the repository is its own commit.
+- The gate came back at exactly 3.4% after that, which was the useful signal: the duplication was somewhere else and had not moved. `sonarcloud.io` is blocked by this environment's egress policy, so the detector was reproduced locally instead — token-based, literals normalised, 100-token minimum, attributing each duplicated block to the lines the PR actually added. It named the block immediately: `api/story/continue.ts` lines 15-55 against both `api/story/generate.ts` and `api/image/generate.ts`, 144 tokens, covering all 18 of this slice's added lines in that file. 18 of 518 new lines is 3.47%, which is the number the gate reported.
+- So the duplication was the fix itself. Giving `/api/story/continue` the correlation id it was missing made its preamble the fourth verbatim copy of one the other three paid POST routes already shared — correlation id, `X-Request-ID`, CORS, method, access control, in that order. The gate was right, and right about the shape of the fix: that sequence is a rule about paid routes rather than a detail of any one of them, and a rule stated four times can be changed in three places. `api/_lib/http/postRoutePreamble.ts` now states it once and all four routes call it; the four handlers lose 133 lines between them and gain 28. `/api/export/save` picks up the 405 log line the other three already wrote, which is the same consistency this slice is about. Reproduced at 0.0% duplication on new code locally, with `tests/api-access-control.test.ts`, `tests/request-correlation-id.test.ts`, and `tests/express-api-routes.test.ts` all still passing.
 - Codex flagged a P2 on the Unicode claim and was right: `[^\p{L}\p{N}]+` reads every combining mark as a separator, so a word is cut apart at each one. `मेरी कहानी` came back as `म-र-कह-न` and `เรื่องของฉัน` as `เร-องของฉ-น` — not shortened but corrupted, the vowel signs and tone marks deleted and hyphens left in their place, which is worse than the fallback it replaced because the result still looks like a slug of the title. `\p{M}` is now retained beside the letters and numbers. Writing the regression test surfaced a second half the P2 did not name: retaining marks does not make `José` typed decomposed and `José` typed precomposed one name, because they are different strings — one story would still download under two names. The title is normalised to `NFC` first. Both are needed: Devanagari, Thai, and Arabic marks do not compose away under `NFC`.
 - Codex also asked for this slice to be split into three, citing the slice-sizing rule. Answered on the PR rather than acted on: that rule is written for `recovery/story-lab-*` slices of unpublished Story Lab work, and this is the "three quick wins" line, whose merged precedent (#237 through #243) is three independent fixes per PR by construction. Splitting was outside what was asked for.
